+++ demo/web/components/popover/main.dart
@@ -1,0 +1,6 @@
+import 'package:angular/angular.dart';
+import 'popover_demo.dart';
+
+main() {
+  bootstrap(PopoverDemo);
+}